Extract all the hash loop setup into `MVM_str_create_loop_state` etc.
This creates a `struct MVM_hash_loop_state` which encapsulates several
repeated local variables. More will be added soon, with more repeated
initialisation code.

`entry_size` is in this struct because it reduces differences between the
analogous functions in the multiple implementations.

Eliminate `MVM_fixkey_hash_code` - use `MVM_str_hash_code` directly for
the bucket index in fixkey hashes. `MVM_fixkey_hash_code` multiplied this
by the golden ratio to mix the hash value, but siphash is itself
cryptographic grade, any subset of bits within it should be suitably mixed
already, and this hash doesn't have iterators or expose its internal order
to external observers.

(This isn't the case for the regular string hash, where iterators can expose
the order, and we need to mix all the bits of the hash table's own salt into
the subset of bits we use for the bucket.)