We need to calculate the hash table's `max_items` in certain assertions.
The insert logic is to set table->max_items to 0 to signal that the next
insert should take action to grow the hash (or make space in the metadata)
before adding any more entries. Hence we can't trust this value to be
"correct" in various assertions. Is *is* correct, just not the value we need.