Replace `if` with `else if` in various hash table scanning loops.
The two `if` tests are for `<` and then `==` on the same pair of values, and
the first if block doesn't alter either, so this is a logical transformation
to make. In theory, an optimising compiler ought to be able to figure this
out, but gcc generates different code (on two different architectures), with
the code from this commit looking slightly tighter, so maybe it does still
need some help. And the human reading the code certainly benefits from the
help.