Eliminate a branch in MVM_*_hash_create_loop_state.
It turns out that we initialise loops on hashes *a lot*. The previous
approach seemed slightly clearer, but cachegrind suggests that we can drop
total instructions dispatched by around 0.1% like this.

Also hoist a constant out of a comparison in the loop in hash delete - it
seems that (at least) gcc can't spot this, and so uses slightly more
instructions. *This* one isn't performance critical :-)