For empty MVMStrHashTable, store only the control structure.
Optimise the storage of freshly created empty `MVMStrHashTable`s by only
storing the control structure, not the control structure plus space for
the initial entries. This way we only use 24 bytes, instead of 144 bytes
(64 bit systems) or 92 bytes (32 bit systems), and grow the hash to that
size if and when we need to store the first entry.

MVMStrHashTable is the only variant where we create many empty hashes (if
any), so the extra complexity of this optimisation isn't useful for the
other 4 hash variants.